# docs/swagger.yaml
definitions:
...
  github_com_mr-isik_gatling-backend_internal_domain.RunConfig:
    properties:
      baseline_run_id:
        type: string
      duration:
        $ref: '#/definitions/time.Duration'
      ramp_down_duration:
        $ref: '#/definitions/time.Duration'
      ramp_up_duration:
        $ref: '#/definitions/time.Duration'
      vus:
        type: integer
    type: object
  github_com_mr-isik_gatling-backend_internal_domain.RunStatus:
    enum:
    - pending
    - running
    - stopping
    - finished
    - failed
    - stopped
    type: string
    x-enum-varnames:
    - StatusPending
    - StatusRunning
    - StatusStopping
    - StatusFinished
    - StatusFailed
    - StatusStopped
...
  github_com_mr-isik_gatling-backend_internal_domain.TestRun:
    properties:
      config:
        $ref: '#/definitions/github_com_mr-isik_gatling-backend_internal_domain.RunConfig'
      created_at:
        type: string
      error_msg:
        type: string
      finished_at:
        type: string
      id:
        type: string
      project_id:
        type: string
      scenario_id:
        type: string
      started_at:
        type: string
      started_by:
        type: string
      status:
        $ref: '#/definitions/github_com_mr-isik_gatling-backend_internal_domain.RunStatus'
    type: object
...
  internal_api_handler.startRunRequest:
    properties:
      config:
        $ref: '#/definitions/github_com_mr-isik_gatling-backend_internal_domain.RunConfig'
      project_id:
        type: string
      scenario_id:
        type: string
    type: object
  time.Duration:
    enum:
    - -9223372036854775808
    - 9223372036854775807
    - 1
    - 1000
    - 1000000
    - 1000000000
    - 60000000000
    - 3600000000000
    format: int64
    type: integer
    x-enum-varnames:
    - minDuration
    - maxDuration
    - Nanosecond
    - Microsecond
    - Millisecond
    - Second
    - Minute
    - Hour
info:
  contact:
    email: support@swagger.io
    name: API Support
    url: http://www.swagger.io/support
  description: AI-Powered Load Testing Tool Backend API.
  termsOfService: http://swagger.io/terms/
  title: Gatling-Backend API
  version: "1.0"
paths:
...
  /v1/runs:
    get:
      description: Lists all test runs for a given project.
      parameters:
      - description: Project ID
        in: query
        name: project_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/github_com_mr-isik_gatling-backend_internal_domain.TestRun'
            type: array
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: List Test Runs
      tags:
      - TestRun
    post:
      consumes:
      - application/json
      description: Starts a new load test execution.
      parameters:
      - description: Start Run Details
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/internal_api_handler.startRunRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Start Test Run
      tags:
      - TestRun
  /v1/runs/{id}:
    get:
      description: Gets detailed information for a specific test run by ID.
      parameters:
      - description: Test Run ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_mr-isik_gatling-backend_internal_domain.TestRun'
        "401":
          description: Unauthorized
          schema:
            additionalProperties: true
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Get Test Run
      tags:
      - TestRun
...
  /v1/runs/{id}/stop:
    post:
      description: Stops a currently running test run.
      parameters:
      - description: Test Run ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Stop Test Run
      tags:
      - TestRun
...
securityDefinitions:
  BearerAuth:
    description: Type "Bearer" followed by a space and JWT token.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"